+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v2.9.1 - 2025/02/20
+
+* [FIX] Updated workspace listing for compatibility with Faraday API. The minimum required Faraday version is now 5.1.0. #25
+* [FIX] Addressed an issue with the vulnerability import functionality. #26
+
+
 # v2.9 - 2022/12/15
 
 * Add create workspace button